feat: 공통 헤더 컴포넌트 및 css 구현
